remove sql_filter_transformer_wrapped
