+++ packages/external-db-mysql/lib/connection_provider.js
@@ -1,7 +1,7 @@
 const mysql = require('mysql')
 const SchemaProvider = require('./mysql_schema_provider')
 const DataProvider  = require('./mysql_data_provider')
-const FilterParser = require('./sql_filter_transformer_wrapped')
+const FilterParser = require('./sql_filter_transformer')
 const DatabaseOperations = require('./mysql_operations')
 
 const init = (cfg, _poolOptions) => {
