+++ packages/external-db-mysql/lib/sql_filter_transformer.spec.js
@@ -1,16 +1,16 @@
 const FilterParser = require('./sql_filter_transformer')
-const { EMPTY_SORT } = require('velo-external-db-commons')
+const { EMPTY_SORT, AdapterOperators, AdapterFunctions } = require('velo-external-db-commons')
 const { Uninitialized, gen } = require('test-commons')
 const { InvalidQuery } = require('velo-external-db-commons').errors
 const each = require('jest-each').default
 const Chance = require('chance')
 const { escapeId } = require('./mysql_utils')
 const chance = Chance()
-
+const { eq, gt, gte, include, lt, lte, ne, string_begins, string_ends, string_contains, and, or, not, urlized } = AdapterOperators //TODO: extract
+const { avg, max, min, sum, count } = AdapterFunctions
 describe('Sql Parser', () => {
     describe('sort parser', () => {
 
-        // todo: should we even check for valid input or should we let the validation library to handle this ?
         test('handles undefined sort', () => {
             expect( env.filterParser.orderBy('') ).toEqual(EMPTY_SORT)
             expect( env.filterParser.orderBy('    ') ).toEqual(EMPTY_SORT)
@@ -74,14 +74,16 @@ describe('Sql Parser', () => {
 
         describe('handle single field operator', () => {
             each([
-                '$ne', '$lt', '$lte', '$gt', '$gte', '$eq',
+                ne, lt, lte, gt, gte, eq,
             ]).test('correctly transform operator [%s]', (o) => {
                 const filter = {
-                    [ctx.fieldName]: { [o]: ctx.fieldValue }
+                    operator: o,
+                    fieldName: ctx.fieldName,
+                    value: ctx.fieldValue
                 }
 
                 expect( env.filterParser.parseFilter(filter) ).toEqual([{
-                    filterExpr: `${escapeId(ctx.fieldName)} ${env.filterParser.veloOperatorToMySqlOperator(o, ctx.fieldValue)} ?`,
+                    filterExpr: `${escapeId(ctx.fieldName)} ${env.filterParser.adapterOperatorToMySqlOperator(o, ctx.fieldValue)} ?`,
                     parameters: [ctx.fieldValue]
                 }])
 
@@ -89,7 +91,9 @@ describe('Sql Parser', () => {
 
             test('correctly extract filter value if value is 0', () => {
                 const filter = {
-                    [ctx.fieldName]: { $eq: 0 }
+                    operator: eq,
+                    fieldName: ctx.fieldName,
+                    value: 0
                 }
 
                 expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -99,10 +103,11 @@ describe('Sql Parser', () => {
 
             })
 
-            // todo: $hasAll ???
-            test('correctly transform operator [$hasSome]', () => {
+            test('correctly transform operator [include]', () => {
                 const filter = {
-                    [ctx.fieldName]: { $hasSome: ctx.fieldListValue }
+                    operator: include,
+                    fieldName: ctx.fieldName,
+                    value: ctx.fieldListValue
                 }
 
                 expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -111,17 +116,21 @@ describe('Sql Parser', () => {
                 }])
             })
 
-            test('operator [$hasSome] with empty list of values will throw an exception', () => {
+            test('operator [include] with empty list of values will throw an exception', () => {
                 const filter = {
-                    [ctx.fieldName]: { $hasSome: [] }
+                    operator: include,
+                    fieldName: ctx.fieldName,
+                    value: []
                 }
 
                 expect( () => env.filterParser.parseFilter(filter) ).toThrow(InvalidQuery)
             })
 
-            test('correctly transform operator [$eq] with null value', () => {
+            test('correctly transform operator [eq] with null value', () => {
                 const filter = {
-                    [ctx.fieldName]: { $eq: undefined } 
+                    operator: eq,
+                    fieldName: ctx.fieldName,
+                    value: undefined                    
                 }
 
                 expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -131,10 +140,12 @@ describe('Sql Parser', () => {
 
             })
 
-            test('correctly transform operator [$eq] with boolean value', () => {
+            test('correctly transform operator [eq] with boolean value', () => {
                 const value = chance.bool()
                 const filter = {
-                    [ctx.fieldName]: { $eq: value } 
+                    operator: eq,
+                    fieldName: ctx.fieldName,
+                    value: value
                 }
 
                 expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -145,9 +156,11 @@ describe('Sql Parser', () => {
             })
 
             describe('handle string operators', () => {
-                test('correctly transform operator [$contains]', () => {
+                test('correctly transform operator [string_contains]', () => {
                     const filter = {
-                        [ctx.fieldName]: { $contains: ctx.fieldValue }
+                        operator: string_contains,
+                        fieldName: ctx.fieldName,
+                        value: ctx.fieldValue
                     }
 
                     expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -157,9 +170,11 @@ describe('Sql Parser', () => {
 
                 })
 
-                test('correctly transform operator [$startsWith]', () => {
+                test('correctly transform operator [string_begins]', () => {
                     const filter = {
-                        [ctx.fieldName]: { $startsWith: ctx.fieldValue }
+                        operator: string_begins,
+                        fieldName: ctx.fieldName,
+                        value: ctx.fieldValue
                     }
 
                     expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -169,9 +184,11 @@ describe('Sql Parser', () => {
 
                 })
 
-                test('correctly transform operator [$endsWith]', () => {
+                test('correctly transform operator [string_ends]', () => {
                     const filter = {
-                        [ctx.fieldName]: { $endsWith: ctx.fieldValue }
+                        operator: string_ends,
+                        fieldName: ctx.fieldName,
+                        value: ctx.fieldValue
                     }
 
                     expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -182,7 +199,9 @@ describe('Sql Parser', () => {
 
                 test('correctly transform operator [$urlized]', () => {
                     const filter = {
-                        [ctx.fieldName]: { $urlized: ctx.fieldListValue } 
+                        operator: urlized,
+                        fieldName: ctx.fieldName,
+                        value: ctx.fieldListValue
                     }
 
                     expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -194,12 +213,13 @@ describe('Sql Parser', () => {
         })
         describe('handle multi field operator', () => {
             each([
-                '$and', '$or'
+                and, or
             ]).test('correctly transform operator [%s]', (o) => {
                 const filter = {
-                    [o]: [ctx.filter, ctx.anotherFilter]
+                    operator: o,
+                    value: [ctx.filter, ctx.anotherFilter]
                 }
-                const op = o === '$and' ? 'AND' : 'OR'
+                const op = o === and ? 'AND' : 'OR'
 
                 expect( env.filterParser.parseFilter(filter) ).toEqual([{
                     filterExpr: `${env.filterParser.parseFilter(ctx.filter)[0].filterExpr} ${op} ${env.filterParser.parseFilter(ctx.anotherFilter)[0].filterExpr}`,
@@ -208,9 +228,10 @@ describe('Sql Parser', () => {
                 }])
             })
 
-            test('correctly transform operator [$not]', () => {
+            test('correctly transform operator [not]', () => {
                 const filter = {
-                    $not: [ ctx.filter ]
+                    operator: not,
+                    value: [ ctx.filter ]
                 }
 
                 expect( env.filterParser.parseFilter(filter) ).toEqual([{
@@ -226,7 +247,7 @@ describe('Sql Parser', () => {
             describe('transform select fields', () => {
                 test('single id field', () => {
                     const aggregation = {
-                        _id: `$${ctx.fieldName}`
+                        projection: [{ name: ctx.fieldName }]
                     }
 
                     expect( env.filterParser.parseAggregation(aggregation) ).toEqual({
@@ -239,10 +260,10 @@ describe('Sql Parser', () => {
 
                 test('multiple id fields', () => {
                     const aggregation = {
-                        _id: {
-                            field1: `$${ctx.fieldName}`,
-                            field2: `$${ctx.anotherFieldName}`
-                        }
+                        projection: [
+                                     { name: ctx.fieldName },
+                                     { name: ctx.anotherFieldName }
+                                    ]
                     }
 
                     expect( env.filterParser.parseAggregation(aggregation) ).toEqual({
@@ -255,15 +276,18 @@ describe('Sql Parser', () => {
 
                 test('process having filter', () => {
                     const aggregation = {
-                        _id: `$${ctx.fieldName}`,
-                        [ctx.moreFieldName]: {
-                            $avg: `$${ctx.anotherFieldName}`
+                        projection: [
+                            { name: ctx.fieldName },
+                            { name: ctx.anotherFieldName, function: avg, alias: ctx.moreFieldName }
+                        ],
+                        postFilter: {
+                            operator: gt,
+                            fieldName: ctx.moreFieldName,
+                            value: ctx.fieldValue
                         }
                     }
 
-                    const havingFilter = {  [ctx.moreFieldName]: { $gt: ctx.fieldValue } }
-
-                    expect( env.filterParser.parseAggregation(aggregation, havingFilter) ).toEqual({
+                    expect( env.filterParser.parseAggregation(aggregation) ).toEqual({
                         fieldsStatement: `${escapeId(ctx.fieldName)}, AVG(${escapeId(ctx.anotherFieldName)}) AS ${escapeId(ctx.moreFieldName)}`,
                         groupByColumns: [ctx.fieldName],
                         havingFilter: `HAVING ${escapeId(ctx.moreFieldName)} > ?`,
@@ -273,16 +297,16 @@ describe('Sql Parser', () => {
 
 
                 each([
-                    ['AVG', '$avg'],
-                    ['MIN', '$min'],
-                    ['MAX', '$max'],
-                    ['SUM', '$sum'],
-                ]).test('translate %s function', (mySqlFunction, wixDataFunction) => {
+                    ['AVG', avg],
+                    ['MIN', min],
+                    ['MAX', max],
+                    ['SUM', sum],
+                ]).test('translate %s function', (mySqlFunction, adapterFunction) => {
                     const aggregation = {
-                        _id: `$${ctx.fieldName}`,
-                        [ctx.moreFieldName]: {
-                            [wixDataFunction]: `$${ctx.anotherFieldName}`
-                        }
+                        projection: [
+                            { name: ctx.fieldName },
+                            { name: ctx.anotherFieldName, function: adapterFunction, alias: ctx.moreFieldName }
+                        ]
                     }
 
                     expect( env.filterParser.parseAggregation(aggregation) ).toEqual({
@@ -295,12 +319,14 @@ describe('Sql Parser', () => {
 
                 test('translate COUNT function', () => {
                     const aggregation = {
-                        _id: `$${ctx.fieldName}`,
-                        count: { $sum: 1 }
+                        projection: [
+                            { name: ctx.fieldName },
+                            { name: '*', alias: ctx.moreFieldName, function: count }
+                        ]
                     }
 
                     expect(env.filterParser.parseAggregation(aggregation)).toEqual({
-                        fieldsStatement: `${escapeId(ctx.fieldName)}, COUNT(${escapeId('*')}) AS ${escapeId('count')}`,
+                        fieldsStatement: `${escapeId(ctx.fieldName)}, COUNT(${escapeId('*')}) AS ${escapeId(ctx.moreFieldName)}`,
                         groupByColumns: [ctx.fieldName],
                         havingFilter: '',
                         parameters: [],
@@ -334,8 +360,8 @@ describe('Sql Parser', () => {
         ctx.fieldValue = chance.word()
         ctx.fieldListValue = [chance.word(), chance.word(), chance.word(), chance.word(), chance.word()]
 
-        ctx.filter = gen.randomFilter()
-        ctx.anotherFilter = gen.randomFilter()
+        ctx.filter = gen.randomWrappedFilter()
+        ctx.anotherFilter = gen.randomWrappedFilter()
     })
 
     beforeAll(function() {
